Added serialization tests for MinimumLogLevel and empty Frequency
